pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-stock-mlops-pipeline"
version = "0.1.0"
description = "End-to-end MLOps pipeline for crypto and stock price prediction"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "numpy>=2.3.4",
    "scikit-learn>=1.7.2",
    "xgboost>=3.1.1",
    "mlflow[databricks]>=3.6.0",
    "dvc[s3]>=3.63.0",
    "evidently>=0.7.16",
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "yfinance>=0.2.40",
    "cryptography>=46.0.0",
    "pyOpenSSL>=25.0.0",
    "ta>=0.11.0",
    "tensorflow-cpu>=2.20.0",
    "optuna>=4.6.0",
    "pre-commit>=4.4.0",
    "pytest-cov>=7.0.0",
    "slowapi>=0.1.9",
    "pip>=26.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.4",
    "pre-commit>=4.4.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
# Python 3.12 target
target-version = "py312"
line-length = 100

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    "build",
    "dist",
    "notebooks",
    "mlruns",
    "catboost_info",
]

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "E402",  # Module level import not at top of file
    "N806",  # Variable in function should be lowercase
    "F841",  # Local variable is assigned to but never used
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = ["S101"]  # Allow assert in tests
"src/models/train_model.py" = ["N806"]
"src/models/tune.py" = ["N806"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
